@@ -4,7 +4,7 @@ version: 1.0.0
 homepage: https://github.com/Skyggedans/rw_speech_recognizer
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.18.4 <3.0.0'
 
 dependencies:
   flutter:
